Better env settings for travis
With env settings on their own lines, Travis will run different tests
for each line. So with four env lines, we get four times as much tests!

Also, if the RELEASE_TESTING is enabled, this will also affect the
cpanm installations. This is not desirable; therefore I export it only
on the 'script' line now.